Refactor /Usage into smaller modules
