+++ crates/chat-cli/src/cli/chat/mod.rs
@@ -3477,7 +3477,7 @@ impl ChatSession {
 
         // Check if context usage indicator is enabled
         let usage_percentage = if ExperimentManager::is_enabled(os, ExperimentName::ContextUsageIndicator) {
-            use crate::cli::chat::cli::usage::get_total_usage_percentage;
+            use crate::cli::chat::cli::usage::usage_data_provider::get_total_usage_percentage;
             get_total_usage_percentage(self, os).await.ok()
         } else {
             None
